+++ src/core/pipeline/items/function/compare.rs
@@ -5,17 +5,17 @@ use async_trait::async_trait;
 use futures_util::future::BoxFuture;
 use crate::core::pipeline::item::Item;
 use crate::core::pipeline::ctx::{Ctx};
-use crate::core::pipeline::items::function::validate::Validity;
+use crate::core::pipeline::items::function::validate::{ValidateResult};
 use crate::core::teon::Value;
 use crate::core::result::Result;
 
-pub trait CompareArgument<T: From<Value> + Send + Sync, O: Into<Validity> + Send + Sync>: Send + Sync {
+pub trait CompareArgument<T: From<Value> + Send + Sync, O: Into<ValidateResult> + Send + Sync>: Send + Sync {
     fn call(&self, old: T, new: T) -> BoxFuture<'static, O>;
 }
 
 impl<T, O, F, Fut> CompareArgument<T, O> for F where
     T: From<Value> + Send + Sync,
-    O: Into<Validity> + Send + Sync,
+    O: Into<ValidateResult> + Send + Sync,
     F: Fn(T, T) -> Fut + Sync + Send,
     Fut: Future<Output = O> + Send + Sync + 'static {
     fn call(&self, old: T, new: T) -> BoxFuture<'static, O> {
@@ -38,7 +38,7 @@ impl<T, O> Debug for CompareItem<T, O> {
 impl<T, O> CompareItem<T, O> {
     pub fn new<F>(f: F) -> CompareItem<T, O> where
         T: From<Value> + Send + Sync,
-        O: Into<Validity> + Send + Sync,
+        O: Into<ValidateResult> + Send + Sync,
         F: CompareArgument<T, O> + 'static {
         return CompareItem {
             callback: Arc::new(f)
@@ -47,7 +47,7 @@ impl<T, O> CompareItem<T, O> {
 }
 
 #[async_trait]
-impl<T: From<Value> + Send + Sync, O: Into<Validity> + Send + Sync> Item for CompareItem<T, O> {
+impl<T: From<Value> + Send + Sync, O: Into<ValidateResult> + Send + Sync> Item for CompareItem<T, O> {
 
     async fn call<'a>(&self, ctx: Ctx<'a>) -> Result<Ctx<'a>> {
         if ctx.get_object()?.is_new() {
@@ -65,11 +65,25 @@ impl<T: From<Value> + Send + Sync, O: Into<Validity> + Send + Sync> Item for Com
             }
             let cb = self.callback.clone();
             let value = cb.call(previous_value.into(), current_value.into()).await;
-            let validity = value.into();
-            if validity.is_valid() {
-                Ok(ctx.clone())
-            } else {
-                Err(ctx.internal_server_error(validity.invalid_reason().unwrap()))
+            let result = value.into();
+            match result {
+                ValidateResult::Validity(validity) => {
+                    if validity.is_valid() {
+                        Ok(ctx.clone())
+                    } else {
+                        Err(ctx.internal_server_error(validity.invalid_reason().unwrap()))
+                    }
+                }
+                ValidateResult::Result(result) => match result {
+                    Ok(validity) => if validity.is_valid() {
+                        Ok(ctx.clone())
+                    } else {
+                        Err(ctx.with_invalid(validity.invalid_reason().unwrap()))
+                    },
+                    Err(error) => {
+                        Err(ctx.unwrap_custom_error(error))
+                    }
+                }
             }
         } else {
             Ok(ctx.clone())