+++ CHANGELOG.md
@@ -221,8 +221,7 @@
 #### 0.0.46
 - Error in pipeline **[DONE]**
 - Rename pipeline modifier to pipeline item **[DONE]**
-- Error in pipeline program functions
-- Custom pipeline function accepts source
+- Error in pipeline program functions **[DONE]**
 - `$alter` item 
 - `@disable`
 - `@action`